Move computation of winner check to the server
