+++ BackEnd/source/server.js
@@ -27,14 +27,183 @@ function updateStateValues(newStateValues) {
   state.gameBoardColumns = newStateValues.gameBoardColumns;
   state.gameBoardRows = newStateValues.gameBoardRows;
 }
-// const cors = require('cors');
+
+function changePlayer(currentPlayerColour) { // PURE FUNCTION
+  if (currentPlayerColour === 'yellow') {
+    return 'red';
+  }
+  return 'yellow';
+}
+
+function placeCounter(columnIndex) {
+  for (let i = state.gameBoardValues[columnIndex].length - 1; i >= 0; i -= 1) {
+    if (state.gameBoardValues[columnIndex][i] == null) {
+      state.gameBoardValues[columnIndex][i] = state.currentPlayer;
+
+      state.currentPlayer = changePlayer(state.currentPlayer);
+
+      const counterPos = `${columnIndex}-${i}`;
+      return counterPos;
+    }
+  }
+  return null;
+}
+
+function checkColumnWin(columnIndex, gameBoardArray) { // PURE FUNCTION
+  let fourInARowCount = 0;
+  for (let i = gameBoardArray[columnIndex].length - 1; i >= 0; i -= 1) { // Loops through each row
+    if ((i + 1) >= 4 && gameBoardArray[columnIndex][i] !== null) {
+      fourInARowCount = 0;
+      for (let j = 1; j < 4; j += 1) {
+        const nextRowUpPiece = gameBoardArray[columnIndex][i - j];
+        if (gameBoardArray[columnIndex][i] !== nextRowUpPiece || nextRowUpPiece === null) {
+          break;
+        } else if (gameBoardArray[columnIndex][i] === nextRowUpPiece) {
+          fourInARowCount += 1;
+        }
+      }
+    } else {
+      break;
+    }
+    if (fourInARowCount >= 3) {
+      return gameBoardArray[columnIndex][i];
+    }
+  }
+  return null;
+}
+
+function checkRowWin(rowIndex, gameBoardArray) { // PURE FUNCTION
+  let fourInARowCount = 0;
+  for (let i = 0; i < gameBoardArray.length - 1; i += 1) {
+    if (gameBoardArray.length - i >= 4) {
+      fourInARowCount = 0;
+      for (let j = 1; j < 4; j += 1) {
+        const nextColumnPiece = gameBoardArray[i + j][rowIndex];
+        if (gameBoardArray[i][rowIndex] !== nextColumnPiece || nextColumnPiece === null) {
+          break;
+        } else if (gameBoardArray[i][rowIndex] === nextColumnPiece) {
+          fourInARowCount += 1;
+        }
+      }
+    } else {
+      break;
+    }
+    if (fourInARowCount >= 3) {
+      return gameBoardArray[i][rowIndex];
+    }
+  }
+  return null;
+}
+
+function checkRightDiagonalWin(colAndRowIndex, gameBoardArray) { // PURE FUNCTION
+  let fourInARowCount = 0;
+  const columnLength = gameBoardArray.length;
+
+  let columnPos = colAndRowIndex[0];
+  let rowPos = colAndRowIndex[1];
+  while (columnPos <= columnLength - 4 && rowPos >= 3) {
+    fourInARowCount = 0;
+    for (let i = 1; i < 4; i += 1) {
+      if (gameBoardArray[columnPos + i][rowPos - i] !== null) {
+        if (gameBoardArray[columnPos][rowPos] === gameBoardArray[columnPos + i][rowPos - i]) {
+          fourInARowCount += 1;
+        } else {
+          break;
+        }
+      } else {
+        break;
+      }
+    }
+    if (fourInARowCount >= 3) {
+      return gameBoardArray[columnPos][rowPos];
+    }
+    columnPos += 1;
+    rowPos -= 1;
+  }
+  return null;
+}
+
+function getLeftDiagnoalPoint(columnIndex, rowIndex, gameBoardArray) { // PURE FUNCTION
+  let columnPos = columnIndex;
+  let rowPos = rowIndex;
+  let lowestLeftDiagonal;
+
+  let foundEnd = false;
+  while (foundEnd === false) {
+    if (columnPos - 1 < 0 || rowPos + 1 > gameBoardArray[0].length - 1) {
+      lowestLeftDiagonal = [columnPos, rowPos];
+      foundEnd = true;
+    }
+    columnPos -= 1;
+    rowPos += 1;
+  }
+  return lowestLeftDiagonal;
+}
+
+function checkLeftDiagonalWin(colAndRowIndex, gameBoardArray) { // PURE FUNCTION
+  let fourInARowCount = 0;
+
+  let columnPos = colAndRowIndex[0];
+  let rowPos = colAndRowIndex[1];
+  while (columnPos >= 3 && rowPos >= 3) {
+    fourInARowCount = 0;
+    for (let i = 1; i < 4; i += 1) {
+      if (gameBoardArray[columnPos - i][rowPos - i] !== null) {
+        if (gameBoardArray[columnPos][rowPos] === gameBoardArray[columnPos - i][rowPos - i]) {
+          fourInARowCount += 1;
+        } else {
+          break;
+        }
+      } else {
+        break;
+      }
+    }
+    if (fourInARowCount >= 3) {
+      return gameBoardArray[columnPos][rowPos];
+    }
+    columnPos -= 1;
+    rowPos -= 1;
+  }
+  return null;
+}
+
+function getRightDiagnoalPoint(columnIndex, rowIndex, gameBoardArray) { // PURE FUNCTION
+  let columnPos = columnIndex;
+  let rowPos = rowIndex;
+  let lowestRightDiagonal;
+
+  let foundEnd = false;
+  while (foundEnd === false) {
+    if (columnPos + 1 > gameBoardArray.length - 1 || rowPos + 1 > gameBoardArray[0].length - 1) {
+      lowestRightDiagonal = [columnPos, rowPos];
+      foundEnd = true;
+    }
+    columnPos += 1;
+    rowPos += 1;
+  }
+  return lowestRightDiagonal;
+}
+
+function checkWinner(columnIndex, rowIndex) {
+  const lowestLeftDiagPoint = getLeftDiagnoalPoint(columnIndex, rowIndex, state.gameBoardValues);
+  const lowestRightDiagPoint = getRightDiagnoalPoint(columnIndex, rowIndex, state.gameBoardValues);
+
+  if (checkColumnWin(columnIndex, state.gameBoardValues) != null) {
+    return checkColumnWin(columnIndex, state.gameBoardValues);
+  } if (checkRowWin(rowIndex, state.gameBoardValues) != null) {
+    return checkRowWin(rowIndex, state.gameBoardValues);
+  } if (checkRightDiagonalWin(lowestLeftDiagPoint, state.gameBoardValues) != null) {
+    return checkRightDiagonalWin(lowestLeftDiagPoint, state.gameBoardValues);
+  } if (checkLeftDiagonalWin(lowestRightDiagPoint, state.gameBoardValues) != null) {
+    return checkLeftDiagonalWin(lowestRightDiagPoint, state.gameBoardValues);
+  }
+  return null;
+}
 
 app.use(express.static('./FrontEnd/source'));
-// app.use(cors());
 app.use(express.json());
 
 app.post('/setup', (req, res) => {
-  // eslint-disable-next-line max-len
   updateStateValues(req.body);
   state.gameBoardValues = setupGameBoardValuesArray(state.gameBoardRows, state.gameBoardColumns);
   res.json({
@@ -43,11 +212,16 @@ app.post('/setup', (req, res) => {
 });
 
 app.post('/placeCounter', (req, res) => {
-  // eslint-disable-next-line max-len
-  updateStateValues(req.body);
-  state.gameBoardValues = setupGameBoardValuesArray(state.gameBoardRows, state.gameBoardColumns);
   res.json({
-    result: state,
+    updatedState: state,
+    placedCounterPos: placeCounter(req.body.column),
+  });
+});
+
+app.post('/checkWin', (req, res) => {
+  const isWin = checkWinner(parseInt(req.body.column), parseInt(req.body.row));
+  res.json({
+    winDetected: isWin,
   });
 });
