+++ FrontEnd/source/board.js
@@ -34,18 +34,6 @@ function resetGame() {
   setupBoard();
 }
 
-function setupGameBoardValuesArray(row, column) { // PURE FUNCTION
-  const customGameBoardValues = [];
-  for (let i = 0; i < column; i += 1) {
-    const tempArray = [];
-    for (let j = 0; j < row; j += 1) {
-      tempArray.push(null);
-    }
-    customGameBoardValues.push(tempArray);
-  }
-  return customGameBoardValues;
-}
-
 function removeHover() {
   for (let i = 0; i < state.gameBoardColumns; i += 1) {
     $(`#${i}`).removeClass('column-hover');
@@ -68,176 +56,49 @@ function showWinner(winningColor) {
   removeHover();
 }
 
-function checkColumnWin(columnIndex, gameBoardArray) { // PURE FUNCTION
-  let fourInARowCount = 0;
-  for (let i = gameBoardArray[columnIndex].length - 1; i >= 0; i -= 1) { // Loops through each row
-    if ((i + 1) >= 4 && gameBoardArray[columnIndex][i] !== null) {
-      fourInARowCount = 0;
-      for (let j = 1; j < 4; j += 1) {
-        const nextRowUpPiece = gameBoardArray[columnIndex][i - j];
-        if (gameBoardArray[columnIndex][i] !== nextRowUpPiece || nextRowUpPiece === null) {
-          break;
-        } else if (gameBoardArray[columnIndex][i] === nextRowUpPiece) {
-          fourInARowCount += 1;
-        }
+function checkWin(columnRowIndex) {
+  const columnRowArray = columnRowIndex.split('-');
+  const body = {
+    column: columnRowArray[0],
+    row: columnRowArray[1],
+  };
+
+  $.ajax({
+    type: 'POST',
+    url: '/checkWin',
+    data: JSON.stringify(body),
+    contentType: 'application/json',
+    success: (result) => {
+      if (result.winDetected != null) {
+        showWinner(result.winDetected);
       }
-    } else {
-      break;
-    }
-    if (fourInARowCount >= 3) {
-      return gameBoardArray[columnIndex][i];
-    }
-  }
-  return null;
-}
-
-function checkRowWin(rowIndex, gameBoardArray) { // PURE FUNCTION
-  let fourInARowCount = 0;
-  for (let i = 0; i < gameBoardArray.length - 1; i += 1) {
-    if (gameBoardArray.length - i >= 4) {
-      fourInARowCount = 0;
-      for (let j = 1; j < 4; j += 1) {
-        const nextColumnPiece = gameBoardArray[i + j][rowIndex];
-        if (gameBoardArray[i][rowIndex] !== nextColumnPiece || nextColumnPiece === null) {
-          break;
-        } else if (gameBoardArray[i][rowIndex] === nextColumnPiece) {
-          fourInARowCount += 1;
-        }
-      }
-    } else {
-      break;
-    }
-    if (fourInARowCount >= 3) {
-      return gameBoardArray[i][rowIndex];
-    }
-  }
-  return null;
-}
-
-function checkRightDiagonalWin(colAndRowIndex, gameBoardArray) { // PURE FUNCTION
-  let fourInARowCount = 0;
-  const columnLength = gameBoardArray.length;
-
-  let columnPos = colAndRowIndex[0];
-  let rowPos = colAndRowIndex[1];
-  while (columnPos <= columnLength - 4 && rowPos >= 3) {
-    fourInARowCount = 0;
-    for (let i = 1; i < 4; i += 1) {
-      if (gameBoardArray[columnPos + i][rowPos - i] !== null) {
-        if (gameBoardArray[columnPos][rowPos] === gameBoardArray[columnPos + i][rowPos - i]) {
-          fourInARowCount += 1;
-        } else {
-          break;
-        }
-      } else {
-        break;
-      }
-    }
-    if (fourInARowCount >= 3) {
-      return gameBoardArray[columnPos][rowPos];
-    }
-    columnPos += 1;
-    rowPos -= 1;
-  }
-  return null;
-}
-
-function getLeftDiagnoalPoint(columnIndex, rowIndex, gameBoardArray) { // PURE FUNCTION
-  let columnPos = columnIndex;
-  let rowPos = rowIndex;
-  let lowestLeftDiagonal;
-
-  let foundEnd = false;
-  while (foundEnd === false) {
-    if (columnPos - 1 < 0 || rowPos + 1 > gameBoardArray[0].length - 1) {
-      lowestLeftDiagonal = [columnPos, rowPos];
-      foundEnd = true;
-    }
-    columnPos -= 1;
-    rowPos += 1;
-  }
-  return lowestLeftDiagonal;
-}
-
-function checkLeftDiagonalWin(colAndRowIndex, gameBoardArray) { // PURE FUNCTION
-  let fourInARowCount = 0;
-
-  let columnPos = colAndRowIndex[0];
-  let rowPos = colAndRowIndex[1];
-  while (columnPos >= 3 && rowPos >= 3) {
-    fourInARowCount = 0;
-    for (let i = 1; i < 4; i += 1) {
-      if (gameBoardArray[columnPos - i][rowPos - i] !== null) {
-        if (gameBoardArray[columnPos][rowPos] === gameBoardArray[columnPos - i][rowPos - i]) {
-          fourInARowCount += 1;
-        } else {
-          break;
-        }
-      } else {
-        break;
-      }
-    }
-    if (fourInARowCount >= 3) {
-      return gameBoardArray[columnPos][rowPos];
-    }
-    columnPos -= 1;
-    rowPos -= 1;
-  }
-  return null;
-}
-
-function getRightDiagnoalPoint(columnIndex, rowIndex, gameBoardArray) { // PURE FUNCTION
-  let columnPos = columnIndex;
-  let rowPos = rowIndex;
-  let lowestRightDiagonal;
-
-  let foundEnd = false;
-  while (foundEnd === false) {
-    if (columnPos + 1 > gameBoardArray.length - 1 || rowPos + 1 > gameBoardArray[0].length - 1) {
-      lowestRightDiagonal = [columnPos, rowPos];
-      foundEnd = true;
-    }
-    columnPos += 1;
-    rowPos += 1;
-  }
-  return lowestRightDiagonal;
-}
-
-function checkWinner(columnIndex, rowIndex) {
-  const lowestLeftDiagPoint = getLeftDiagnoalPoint(columnIndex, rowIndex, state.gameBoardValues);
-  const lowestRightDiagPoint = getRightDiagnoalPoint(columnIndex, rowIndex, state.gameBoardValues);
-
-  if (checkColumnWin(columnIndex, state.gameBoardValues) != null) {
-    showWinner(checkColumnWin(columnIndex, state.gameBoardValues));
-  } else if (checkRowWin(rowIndex, state.gameBoardValues) != null) {
-    showWinner(checkRowWin(rowIndex, state.gameBoardValues));
-  } else if (checkRightDiagonalWin(lowestLeftDiagPoint, state.gameBoardValues) != null) {
-    showWinner(checkRightDiagonalWin(lowestLeftDiagPoint, state.gameBoardValues));
-  } else if (checkLeftDiagonalWin(lowestRightDiagPoint, state.gameBoardValues) != null) {
-    showWinner(checkLeftDiagonalWin(lowestRightDiagPoint, state.gameBoardValues));
-  }
-}
-
-function changePlayer(currentPlayerColour) { // PURE FUNCTION
-  if (currentPlayerColour === 'yellow') {
-    return 'red';
-  }
-  return 'yellow';
+    },
+  });
 }
 
 function placeCounter(columnIndex) {
-  for (let i = state.gameBoardValues[columnIndex].length - 1; i >= 0; i -= 1) {
-    if (state.gameBoardValues[columnIndex][i] == null) {
-      state.gameBoardValues[columnIndex][i] = state.currentPlayer;
-      const counterPos = `${columnIndex}-${i}`;
-      $(`#${counterPos}`).css('background-color', state.currentPlayer);
+  let newlyPlacedPiece;
+  const body = {
+    column: columnIndex,
+  };
+  $.ajax({
+    type: 'POST',
+    url: '/placeCounter',
+    data: JSON.stringify(body),
+    contentType: 'application/json',
+    success: (result) => {
+      const resultObject = result.updatedState;
+      newlyPlacedPiece = result.placedCounterPos;
 
-      checkWinner(columnIndex, i);
-      break;
-    }
-  }
-  state.currentPlayer = changePlayer(state.currentPlayer);
-  $('#player-counter').css('background-color', state.currentPlayer);
+      $(`#${newlyPlacedPiece}`).css('background-color', state.currentPlayer);
+
+      updateStateValues(resultObject);
+
+      $('#player-counter').css('background-color', state.currentPlayer);
+
+      checkWin(newlyPlacedPiece);
+    },
+  });
 }
 
 function drawBoardToHtml(columnSize, rowSize) {
@@ -266,12 +127,6 @@ function displayInGameAssets() {
   $('#reset-button').attr('style', 'display: block');
 }
 
-function testSetupBoard() {
-  $.get('http://localhost:8080/setup', () => {
-
-  });
-}
-
 // eslint-disable-next-line no-unused-vars
 function setupBoard() {
   resetBoard();
@@ -295,7 +150,6 @@ function setupBoard() {
     },
   });
 }
-
 // module = module || {};
 // if (typeof module !== 'undefined') {
 //   module.exports = {
